+++ src/main/java/dev/ultima/entityquery/SectionCounterAccess.java
@@ -1,0 +1,5 @@
+package dev.ultima.entityquery;
+
+public interface SectionCounterAccess {
+    EntitySectionCounters ultima$counters();
+}